requirements.txt
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-Classful==0.16.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
marshmallow-sqlalchemy==1.1.0
jsonschema==4.23.0
python-dotenv==1.0.1